requirements.txt
picamera
RPi.GPIO==0.6.3
